requirements.txt
altgraph==0.17.4
et_xmlfile==2.0.0
lxml==6.0.0
numpy==2.2.3
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
packaging==24.2
pefile==2023.2.7
php==1.2.1
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.2
PyMuPDF==1.26.3
pypdf==5.1.0
PyPDF2==3.0.1
pywin32-ctypes==0.2.3
setuptools==78.1.0
tk==0.1.0
python-dotenv==1.1.1

